requirements.txt
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
matplotlib>=3.5.0
seaborn>=0.11.0
gensim>=4.2.0
bertopic>=0.15.0
sentence-transformers>=2.2.0
umap-learn>=0.5.0
hdbscan>=0.8.0
pymorphy2>=0.9.0
nltk>=3.7.0
wordcloud>=1.9.0
plotly>=5.0.0
jupyter>=1.0.0 